Use `/auth` as Slack redirect url
* Install Reach Router
* Split existing `App` code between `Home` and `AuthCodeReceiver`
    * `AuthCodeReceiver` just receives `code` from query param and logs in
    * `Home` does the rest
